fix(ci): landing page auth redirect + trip creation navigation race condition
