+++ apps/web/tests/e2e/helpers/trips.ts
@@ -32,7 +32,7 @@ export async function createTrip(
   await pickDate(page, tripDetail.endDateButton, endDate);
   await tripDetail.continueButton.click();
   await expect(tripDetail.step2Indicator).toBeVisible();
-  await tripDetail.createTripButton.click();
+  await tripDetail.createTripButton.click({ noWaitAfter: true });
   // Step 3: timezone confirmation — click "Go to trip" to complete navigation
   await expect(tripDetail.goToTripButton).toBeVisible({ timeout: NAVIGATION_TIMEOUT });
   await tripDetail.goToTripButton.click();
